+++ packages/terra-search-field/tests/nightwatch/DelayedSearchField.jsx
@@ -1,5 +1,8 @@
 import React from 'react';
+import Base from 'terra-base';
 import SearchField from '../../lib/SearchField';
+
+const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
 
 class DelayedSearchField extends React.Component {
 
@@ -13,12 +16,12 @@ class DelayedSearchField extends React.Component {
 
   render() {
     return (
-      <div>
+      <Base locale={locale}>
         <div id="search-callback-text">
           Search Text: {this.state.searchText}
         </div>
         <SearchField searchDelay={1000} onSearch={(searchText) => { this.setState({ searchText }); }} />
-      </div>
+      </Base>
     );
   }