+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Added
+*Added disabled prop.
+
+### Changed
+* Uplift to Button 2.0
 
 1.20.0 - (February 1, 2018)
 ------------------